# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateServiceInput:
    properties:
      branchId:
        type: string
      description:
        type: string
      durationMins:
        type: integer
      name:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: tangtnp@example.com
    name: Tang Tanathorn
  description: |-
    QueuePlus backend API for queue management system.
    This API provides authentication, branch management, service management, and queue operations.
  license:
    name: MIT
  title: QueuePlus API
  version: "1.0"
paths:
...
  /services:
    get:
      description: Get all non-deleted services, optionally filtered by branchId
      parameters:
      - description: Branch ID
        in: query
        name: branchId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: List services
      tags:
      - Services
    post:
      consumes:
      - application/json
      description: Create a new service under a branch
      parameters:
      - description: Service payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateServiceInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create service
      tags:
      - Services
  /services/{id}:
    delete:
      description: Soft delete service by setting isDeleted=true
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Soft delete service
      tags:
      - Services
    get:
      description: Get a single non-deleted service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get service by ID
      tags:
      - Services
    put:
      consumes:
      - application/json
      description: Update a non-deleted service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      - description: Service payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateServiceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update service
      tags:
      - Services
  /services/{id}/hard:
    delete:
      description: Permanently delete service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Hard delete service
      tags:
      - Services
  /services/search/by-tags:
    get:
      description: Search non-deleted services using MongoDB $all operator on tags
      parameters:
      - collectionFormat: multi
        description: Tags
        in: query
        items:
          type: string
        name: tags
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Search services by tags
      tags:
      - Services
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"